HUM-304 Initial AIO build for Ceph using Ceph-Ansible
This is an initial AIO build of Ceph using Ceph-Ansible.
The default build includes LXC containers as follows:

2 dedicated mon hosts
2 dedicated osd hosts
1 colocated osd/mon/rgw host

rpc-maas integration is ready and built, but is currently disabled as it
is failing the gate - we'll need to investigate that issue.

Additionally, due to https://github.com/ceph/ceph-ansible/issues/1921 we
are not running multiple RGW hosts for testing. Once we move to
ceph-ansible v3.0.0, once it is released, we should re-enable additional
RGW hosts.

This repo reuses existing roles for creating lxc containers, and aims to
provide a very thin layer that will allow us to control and test the
versions of ceph-ansible that we deploy and use.

The end goal is to enable us to deploy Ceph standalone using the
tested rpc-ceph repository which is just an import of ceph-ansible, a
set of group_vars with sensible defaults, and a set of playbooks to
use.