Mirror group_vars for tests and playbooks
We should have a sane set of default values we test against, and not a
list of group_vars for testing that aren't mirrored to the playbooks
directory.

Additionally, this cleans up some of the vars and separates them into
the correct places for clarity.

Finally this removes some redundant tasks in the ceph-storage setup
playbook and utilized the built in options in ceph-ansible itself.