Set default journal size to 20GB
The journal size should be 2 * expected throughput * filestore max sync
interval. Our expected throughput on our best non-SSD drives is 140MB/s.
filestore max interval has a default of 5. That would mean a journal
size of 14GB. Let's round that up to 20GB as a lazy and safe estimate,
for a start point for journal devices.