Clean up rsyslog logs, and create cleanup.yml
When re-running the playbooks the test for logs will fail because there
is a good chance the osds are not on the same hosts - meaning additional
logs are created.

To fix this we should remove anything in /var/log/rsyslog/ on a new run.
Additionally, we can remove the "mount" cleanup since we no longer mount
the partitions - this is a relic from the initial implementation.
Finally, we should move the cleanup tasks into a separate file for
clarity.