Allow the size of the osd/journal devices to be set
Using the rpc_ceph_test_osd_size var we can now set the size of the osd
and journal devices.
We setup 4 per host, and there are 3 OSD hosts, so that will mean we
need 12x rpc_ceph_test_osd_size spare on the AIO localhost.

In order to use this we will need to ensure the osd_journal_size is
rpc_ceph_test_osd_size/3 otherwise the journals will be too big for the
journal partition.