Reduce size of AIO build
We were building 3 OSDs per OSD host at 10GB each + a journal, this
equated to 120GB (40/host), which is quite a storage footprint for an
AIO.

We don't need to build that many OSDs, and we can also reduce the size
of each disk used, as well as the journal. By dropping to 2 drives per
host + journal, and setting the size to 3GB each, we can reduce this to
27GB instead of 120.

If we need further space we could consider removing one of the OSD
hosts.