[WIP] Move to use lxc-device add
ceph-ansible added a gpt partition change, which can't be done without
the appropriate permissions.

This patch moves to use lxc-device add instead of doing mknod manually
and partitioning the device before hand.

Additionally, trim the osd hosts to 2, there is no quorum so simply
testing that the osd service will work on a host by itself or on a host
with other services should suffice.