Move to latest ceph-ansible v3.0 branch
In order to adopt 3.0.15+, and newer versions of ceph a few changes need
to take place:

* We need to ensure gpt is set on the partition as well as on the loop
device.
* We need to mknod both /dev/sda and /dev/sda1 as the partition we have
created. This is a hacky work around, since ceph-disk doesn't see the
loopback device as a disk or a partition, but ceph-ansible requires a
disk with a partition at <dev_name>1. This will change in the future
since the recommended approach is moving to use ceph-volume instead of
ceph-disk.
* The osd_journal_size has changed to just journal_size, we can support
this by setting journal_size to default to osd_journal_size.
* A bug fix to the radosgw systemd scripts has merged so we can remove
unnecessary tasks now.

This pins the version of ceph-ansible to the head of the stable 3.0
branch, pending a new release (3.0.16), this should do the trick!