[CEPHSTORA-51] Use maas_use_api for Ceph tests
We don't use the MaaS api which is giving us some fake results when
monitoring would fail, but is showing up as passing still.

We can fix this by enabling it and specifying the maas_entity_name to be
the ansible_hostname rather than the inventory_hostname for tests since
localhost will not show up as an entity.

These settings are all setup in /etc/openstack_deploy/user_*_vars.yml
for the OpenStack installs, but are not configured for MaaS at all.