Fix cleanup of scbench pool after test-rgw
The scbench pool cleanup needs the mon_allow_pool_delete setting to be
true to work properly. We can set this on the fly, if it isn't set, and
ensure it is set again before we clean it up.

Longer term we may want to move this kind of bench to the fiobench
tasks, and use the benchmarking as a proof that Ceph is working
properly.