[CEPHSTORA-68] Use correct version of Ansible for ceph-ansible
This patch creates a wrapper "ceph-ansible" that ensures that the
correct Ansible version (and dependencies) can be installed for
ceph-ansible specifically are used.

Additionally, this removes the installation of tox and bindep which
aren't required. Documentation has been updated to reflect this. This
should make it easier for operators since they won't have to load into
the specific venv and can call the correct version using "ceph-ansible"
instead.