[CEPHSTORA-86] Allow defaults to be overridden
We need to set the default pg num and pgp num down from 128 for testing,
128 creates pools that are too big and causes ceph radosgw to not start.

To do this properly this patch sets up a "ceph_conf_overrides_extra"
variable that is combined with ceph_conf_overrides_all to allow
additional settings to be specified that will overwrite the existing
ones in group_vars, but still use settings that aren't specifically
overwritten.

This saves the deployer having to completely respecify the var including
all defaults that they may not want to change at all.