Pause before setting up logging
The logging role doesn't continue to update logs that are shipped, this
causes issues when the role is run but services haven't started properly
yet.

This PR adds a pause in that will wait for 10 seconds to allow the
services some time to start up.

This should hopefully alleviate some of the logging test isues we've
seen in the gate.