Rework use cases and change service name
Did away with the rgw and keystone only use case because it ended
being identical to the the rgw, keystone and openstack use case.
Renamed the first use case rgw without auth system and it still
requires rpc-ceph variables.

Changed the service name to "cephrgw" and the service description
to "Ceph Object Store"