RE-1276 Fix testing rpc-ceph with rpc-maas
rpc-maas uses rpc-ceph to test the code related to monitoring a Ceph
deployment. rpc-ceph uses rpc-maas to test that that MaaS monitoring
using the master branch works. When testing the repository rpc-maas, the
MaaS tests run by rpc-ceph need to be disabled. IRR_CONTEXT has been
used to determine if the tests should be run however that is broken in
`gating/pre_merge_test/run` due to it being set to RE_JOB_BRANCH which
is only used in post-merge jobs and because it can't be overridden.

This change fixes the use of IRR_CONTEXT so that it can be set in
rpc-maas and adds a new variable, TEST_RPC_MAAS, that is more
descriptive with regards to its purpose. Once this is merged and
rpc-maas has switched to using TEST_RPC_MAAS, IRR_CONTEXT can be
removed.