CEPHSTORA-105 Setup HAProxy outside of testing
We should setup haproxy when the haproxy_all host group is specified,
and when rados gw hosts exist. This shouldn't be determined by the
OpenStack integration, and Ceph should manage it's own VIP.

This PR moves the playbook into "playbooks" dir and makes adjustments to
the role-requirements, group_vars defaults, README and base inventory
for testing.

Additionally, to ensure roles/plugins for the Ceph roles, that would
overlap with any existing deploys - we should name space the roles,
e.g. use "ceph_roles". This will mean that issues don't arise as a
result of other deployments using the same roles and overwriting the
existing Ceph roles.