[WIP] nSpawn container test
The lxc containers we're using make attaching drives that will work with
Ceph hard.

This is an initial exploration into using nSpawn containers which are
closer to a "chroot" and should allow us to setup the storage easier for
Ceph purposes.

Currently completed: nSpawn containers with networking and appropriate
ips configured.
Pending: Setup the drives to the nspawn containers.