Mods after review
Changed the benchmark defaults to be more in line with something
used on production system

Renamed the user and subuser accounts to be more in line with
an rgw benchmark

Change the hummingbird url to be an ansible variable instead of
being hardcoded.  Also changed the subaccount password to use a
variable instead of being hardcoded.

Updated documenation to reflect these changes and added an
explanation about setting the values in tests/test-vars.yml to
be AIO specific while the defaults are more for a production
environment.

Could not change the user and subuser creation to the rgws nodes
because there is no ceph.admin.key