CEPHSTORA-131 Add RPC-O Integration gate
Add an rpc-o integration "rpco_newton" scenario.
This builds a base-kit from the newton-rc branch, using only the
following services: Nova, Glance, Keystone, Cinder, Neutron.

In order to build this in a one time go without having to redeploy
certain services for config updates, we deploy in the following order:

* Bootstrap RPC-O ansible
* Bootstrap RPC-O AIO
* Replace env.d/cinder.yml to run is_metal: False
* Deploy RPC-O up until Keystone
* Add configuration options for Ceph
* Deploy Ceph
* Add configuration options for RPC-O
* Deploy Rest of RPC-O (Glance, Cinder, Nova, Neutron)
* Deploy tempest, then test deployment using tempest