CEPHSTORA-151 Adjust run script to work with perf2-15
The RPC-O AIO runs on perf2-15 because the general1-8 does not suffice.
We should adjust the RPC-Ceph gate to be able to do the same, so that
the integrated AIO can work more reliably.

In order for this to happen we need to adjust how we pass the values for
the BOOTSTRAP_OPTS and just use the vars we already had specified.