CEPHSTORA-155 Let regular AIO run in perf2-15
In #118 we fixed the integrated build to work against perf2-15 flavor,
to match what is used by the RPC-O gate.

This fixes the regular rpc-ceph AIO to work on perf2-15, by copying the
approach taken by OSA. This will allow us to adopt the perf2-15 if we
run into performance/memory issues in the future.