CEPHSTORA-158 Only run benchmark tests during Keystone integration
To reduce the runtime of the functional test, we can run the
Benchmark tests against the Keystone/RGW integration test only. This
test is already shorter because it does not run the MaaS installation
steps.

In this way we reduce the length of time for the functional and RPC
integration tests and still ensure that benchmarking is tested.