CEPHSTORA-171 Default bootstrap_host_data_disk_device_force
In the integrated build RPC-O has already setup the disk partitions,
this means that it won't run unless
bootstrap_host_data_disk_device_force is set to True, since it isn't set
at all the task fails.

This PR defaults to false, which will work for the RPC-O integration
build and the normal rpc-ceph builds on perf2-15 instances.