CEPHSTORA-199 Fix disable_transparent_hugepage
We should default this to the ceph-ansible default which is "true". This
was set to false when the initial AIO was setup, but should not have
been left in place.

In order for that setting to work we need to configure OSD hosts to
remounte /sys rw for the AIO containers.