CEPHSTORA-222 Use Ansible 2.5.2.0
This PR reorganizes how we adopt plugins to facilitate ansible 2.5.2.0
Primarly, we no longer need all the plugins from OSA for Ceph to work,
since the config_template plugin has moved. To address this the PR
moves the osa-plugins repo dependency to the tests role requirements,
and installs a separate config_template role dependency.

This allows us to use ceph-ansible v3.0.x with Ansible 2.5.

Additionally, as part of getting ready for the release of ceph-ansible
v3.1.0 we should adopt the paths for the ceph-ansible library, callback
and actions plugins.