CEPHSTORA-222 Use Ansible 2.5.7.0 and ceph-ansible 3.1
This PR changes to use Ansible 2.5.7.0 (the latest Ansible 2.5 version)
along with ceph-ansible 3.1

To get this working we clean up some unrequired references in
ansible.cfg, adjust the bootstrap-ansible.sh file and adjust the plugins
and ceph-ansible role versions.

The master branch of OSA-plugins includes a connection module that will
work with ansible 2.5+, whereas stable/queens does not. Additionally,
the config_template module required for ceph-ansible has been updated in
ceph-ansible stable-3.1 so we no longer need the plugins directory for
prod and should move it to the testing roles to clone.

The openstack_hosts role also has some changes that are required for
Ansible 2.5 to run in our keystone_rgw scenario test. This is a testing
only role so we can bump this to master safely.
8f7114732df71249a907fa738da3c6b40492b50d is required for
openstack_hosts to work with Ansible 2.5+

Finally, this PR adjusts the update_dependency job so that v3.1 of
ceph-ansible is used as the base, and skips the
openstack-ansible-plugins repo.