[WIP] Initial PR to add bionic AIO support
In order for Bionic to work at this point, we need the following:

Utilise the ceph_origin 'distro' until download.ceph.com has packages
for Bionic.

Set the lxc.apparmor.profile setting instead of the lxc.aa_profile
setting, this changed in lxc version 3 (which is default in bionic).

Move the plugins into a role dir so it can be imported as a dependency
for upstream roles.

TODO: Update the roles - these need to be setup for bionic support, so
for OSA roles that means bumping to master - this doesn't include
keystone and galera/rabbitmq etc. which would only be needed in
keystone_rgw job. I've tested this by setting them to master, just to
release this as a "you can start messing about with Bionic".