Remove doc files, Rework README.md
The projects doc files are being moved to the
directory "rpc-ceph-internal" in the
repo "https://github.com/rcbops/privatecloud-docs" which is handled
by the Rackspace Document Team. We will no longer keep any
documentation in our repo.

Reworked the README.md file to give instruction on how to view
the new documentation and things to do if you want add or
modify project docs.