CEPHSTORA-191 Fix versions of rsyslog/haproxy etc
We shouldn't use head of branch for a specific release. We should
instead use the released versions that will remain stable.

Additionally, ceph-ansible carries all the plugins it requires, we
should utilise those instead of getting the openstack-ansible-plugins
directory.

Finally, in moving to newer versions of the upstream OSA roles (used in
testing), we need to adjust the way the pip_install role is called. This
used to be a meta/main.yml dependency for roles but now needs to be
explicitly called, which happens in the openstack_hosts role.