CEPHSTORA-321 Add OSA plugins as a base role
The OSA plugins are required to perform keystone integration.
Another solution would be to utilize the Shade modules, and use the
openstack_openrc role to add a default clouds.yaml. However, until we
use Ansible 2.5.0 the os_keystone_endpoint module is not available, and
we wouldn't be able to create the endpoints. As such this is the
simplest approach to fixing up our integration jobs and ensuring we can
create keystone endpoints/users/roles/services.