CEPHSTORA-323 Fix rpco_integration for Queens+
In Queens release of OSA the inventory directory moved from the
playbooks dir to the top level of the OSA directory.

Additionally, the haproxy user_config moved into it's own scenario file
which needs to be included.

Finally, the data-disk creation changed to use btrfs for /var/lib/lxc
but we shouldn't try mount it if it already exists in anycase.