CEPHSTORA-325 Specify cpu states to adjust and allow enabling
This PR adds 2 variables, "enable_cpuidle_states" and
"disable_cpuidle_states", which will allow the operator to specify the
states to disable or enable.

Additionally, adds a regexp to ensure that multiple lines aren't added
to the same file, and that a specific state is either enabled or
disabled, but not both.

NB if you specify the same values in both disable or enable it will
enable them. (Enabling takes precedence).