CEPHSTORA-330 Automatically update all Ansible roles
We recently added an automated job to update the deps for ceph-ansible
automatically. This PR includes updates to all other roles that we use
for both testing and production.

This PR also adjusts the way the ansible-role-requirements-editor.py
functions, meaning that you can specify the repo to update by its URL,
if no name is present it will take that as the identified, allowing us
to update roles such as the plugins role, which does not have a name and
src that are the same.

Finally, instead of cloning the ceph-ansible repo and figuring out the
tag that way, we can utilise git ls-remote instead.