Turning the rpc-maas tests off by default
There seems to be an issue with the master branch of the repo used in
the file tests/ansible-role-test-requirements.yml for
rpc-maas/tests/common.  Using a previous branch seems to work ok but
It's not a good idea for rpc-ceph to pin it's rpc-maas testing to a
specific branch.  The working branch is:

  version: b146d649e675d748a58af238c7b37138b8194f1f

name: rpc-maas/tests/common
  scm: git
  src: https://git.openstack.org/openstack/openstack-ansible-tests
  version: master