JIRA: CEPHSTORA-499 Fix hanging aio jobs
When setting up the storage for aio devices the osds
(osd1, osd2, allsvc), the losetup task has a tendancy
to hang.  May tasks in this play are delegating to the
physical host and running disk related tasks in parallel.
The losetup task seems to hang often when this happens
resulting in failed builds due time timeouts on the task.
This change is to set up the play to run things in serial
to cut down on conflicts.